Added separated libraries
