+++ web/themes/custom/caesar/.postcssrc.js
@@ -17,9 +17,8 @@ module.exports = {
     require('postcss-pxtorem')({
       propList: ['*'],
     }),
-    // require('stylelint')({
-    //   configFile: __dirname + '/caesar.breakpoints.yml',
-    //   fix: true,
-    // }),
+    require('stylelint')({
+      fix: true,
+    }),
   ]
 }
